// packages/monaco-lsp-client/src/types.ts
export enum MarkupKind {
    /**
     * Plain text is supported as a content format
     */
    PlainText = 'plaintext',
    /**
     * Markdown is supported as a content format
     */
    Markdown = 'markdown'
}
// ...
export type DocumentSelector = (DocumentFilter)[];

export type ProgressToken = number | string;
// ...
export type MarkedString = string | {
    language: string;
    value: string
};
// ...
export interface WorkDoneProgressOptions {
    workDoneProgress?: boolean;
}

/**
 * General text document registration options.
 */
export interface TextDocumentRegistrationOptions {
    /**
     * A document selector to identify the scope of the registration. If set to null
 * the document selector provided on the client side will be used.
     */
    documentSelector: DocumentSelector | null;
}
// ...
export interface HoverParams extends TextDocumentPositionParams, WorkDoneProgressParams {
}

/**
 * The result of a hover request.
 */
export interface Hover {
    /**
     * The hover's content
     */
    contents: MarkupContent | MarkedString | (MarkedString)[];
    /**
     * An optional range inside the text document that is used to
 * visualize the hover, e.g. by changing the background color.
     */
    range?: Range;
}

/**
 * Registration options for a {@link HoverRequest}.
 */
export interface HoverRegistrationOptions extends TextDocumentRegistrationOptions, HoverOptions {
}
// ...
export interface TextDocumentPositionParams {
    /**
     * The text document.
     */
    textDocument: TextDocumentIdentifier;
    /**
     * The position inside the text document.
     */
    position: Position;
}

export interface WorkDoneProgressParams {
    /**
     * An optional token that a server can use to report work done progress.
     */
    workDoneToken?: ProgressToken;
}
// ...
export interface Range {
    /**
     * The range's start position.
     */
    start: Position;
    /**
     * The range's end position.
     */
    end: Position;
}
// ...
export interface TextDocumentIdentifier {
    /**
     * The text document's uri.
     */
    uri: string;
}
// ...
export interface Position {
    /**
     * Line position in a document (zero-based).
 *
 * If a line number is greater than the number of lines in a document, it defaults back to the number of lines in the document.
 * If a line number is negative, it defaults to 0.
     */
    line: number;
    /**
     * Character offset on a line in a document (zero-based).
 *
 * The meaning of this offset is determined by the negotiated
 * `PositionEncodingKind`.
 *
 * If the character value is greater than the line length it defaults back to the
 * line length.
     */
    character: number;
}
// ...
export interface MarkupContent {
    /**
     * The type of the Markup
     */
    kind: MarkupKind;
    /**
     * The content itself
     */
    value: string;
}
// ...
export interface HoverOptions extends WorkDoneProgressOptions {
}